Move authentication logic into separate package
